agents/drone/requirements.txt
#paho-mqtt==1.6.1
#pyjwt==2.9.0
paho-mqtt
pyjwt